Fix a 64-bit FileErase() crash

+++ Misc/ZipArchive.cpp
@@ -1959,7 +1959,7 @@ bool ZipArchive::FileErase(const char* fileName)
 
 		size_t sizeToRemove = m_fileEntryOffsets[index + 1] - m_fileEntryOffsets[index];
 		memcpy(GetFileEntry(index), GetFileEntry(index + 1), m_fileEntriesSizeBytes - m_fileEntryOffsets[index + 1]);
-        memcpy(m_fileEntryOffsets + index, m_fileEntryOffsets + index + 1, (m_fileEntryCount - index - 1) * sizeof(unsigned int));
+        memcpy(m_fileEntryOffsets + index, m_fileEntryOffsets + index + 1, (m_fileEntryCount - index - 1) * sizeof(size_t));
 		m_fileEntryCount--;
 
 		// Reinsert all following entries into the map.
